+++ apps/anchor-demo/jest.setup.ts
@@ -1,1 +1,1 @@
-import './src/__tests__/__mocks__/nativeModules';
+import './src/__tests__/__testboundaries__/nativeBoundaries';